requirements.txt
langchain>=0.1.0
langchain-text-splitters>=0.0.1
langchain-community>=0.0.10
langchain-core>=0.0.10
langchain-mistralai>=0.0.2
PyPDF2>=3.0.1
python-dotenv>=1.0.0
streamlit>=1.18.1
faiss-cpu>=1.7.4
altair<5
tiktoken>=0.4.0
requests>=2.28.0
huggingface-hub>=0.14.1